Adjust position of oscillator headers and nearby cap.
This is to allow a crystal with 200 mil leads to be soldered, while
leaving the headers unpopulated.

There might not quite be enough room to solder a 14-dip oscillator in
place of the other two headers, but I don't think I would find that
useful.

Also tweak a few vias to avoid intersecting with the silkscreen.